LayoutTests/imported/w3c/web-platform-tests/css/cssom-view/client-props-input-expected.txt



FAIL client* on input / textarea assert_equals: INPUT : clientWidth should be the padding box for textarea, content box for input expected 300 but got 304
